requirements.txt
pyautogui ~= 0.9.53
pydirectinput ~= 1.0.4
pillow ~= 8.3.2
opencv-python ~= 4.5.3.56
numpy ~= 1.21.2